third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeysession-reference.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeySession lifetime without release()</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around
            //       AND the session has not received a close() event)

            promise_test(function(test)
            {
                var mediaKeys;
                var mediaKeySession1;
                var mediaKeySession2;
                var mediaKeySession3;
                var initDataType;
                var initData;

                // Even though there should be no existing objects, start by
                // running gc() and verifying that none exist. This also
                // allows the failures to be reported from inside a promise
                // so that the test fails properly.
                return createGCPromise().then(function() {
                    verifyMediaKeyAndMediaKeySessionCount(0, 0, 'After initial gc()');

                    return navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration());
                }).then(function(access) {
                    initDataType = access.getConfiguration().initDataTypes[0];
                    initData = getInitData(initDataType);
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;
                    assert_equals(typeof mediaKeys.createSession, 'function');
                    verifyMediaKeyAndMediaKeySessionCount(1, 0, 'MediaKeys.create()');

                    // Create 3 sessions.
                    mediaKeySession1 = mediaKeys.createSession();
                    return mediaKeySession1.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession1.sessionId && mediaKeySession1.sessionId.length > 0);
                    verifyMediaKeyAndMediaKeySessionCount(1, 1, 'MediaKeys.createSession(1)');

                    mediaKeySession2 = mediaKeys.createSession();
                    return mediaKeySession2.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession2.sessionId && mediaKeySession2.sessionId.length > 0);
                    verifyMediaKeyAndMediaKeySessionCount(1, 2, 'mediaKeys.createSession(2)');

                    mediaKeySession3 = mediaKeys.createSession();
                    return mediaKeySession3.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession3.sessionId && mediaKeySession3.sessionId.length > 0);
                    verifyMediaKeyAndMediaKeySessionCount(1, 3, 'mediaKeys.createSession(3)');

                    // Run gc(). All sessions should remain as we have a
                    // reference to each one. However, running gc()
                    // asynchronously should free up the last PromiseResolver.
                    return createGCPromise();
                }).then(function(result) {
                    verifyMediaKeyAndMediaKeySessionCount(1, 3, 'After gc()');

                    // Now drop references to 2 of the sessions. Even though we
                    // don't have a reference, MediaKeys is still around (and
                    // the sessions aren't closed), so the objects won't be
                    // collected.
                    mediaKeySession1 = null;
                    mediaKeySession2 = null;
                    return createGCPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    verifyMediaKeyAndMediaKeySessionCount(1, 3, 'After second gc()');

                    // Now drop the reference to MediaKeys. It and the 2
                    // unreferenced sessions should be collected. Since
                    // MediaKeySessions remain alive as long as MediaKeys is
                    // around, it is possible that gc() checks one or both
                    // MediaKeySession objects first, and doesn't collect them
                    // since MediaKeys hasn't been collected yet. Thus run gc()
                    // twice, to ensure that the unreferenced MediaKeySession
                    // objects get collected.
                    mediaKeys = null;
                    return createGCPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    verifyMediaKeyAndMediaKeySessionCount(0, 1, 'After mediaKeys = null');

                    // Drop the reference to the last session. It should get
                    // collected now since MediaKeys is gone.
                    mediaKeySession3 = null;
                    return createGCPromise();
                }).then(function(result) {
                    verifyMediaKeyAndMediaKeySessionCount(0, 0, 'After final gc()');
                });
            }, 'MediaKeySession lifetime without release()');
        </script>
    </body>
</html>
